+++ Flashcards.wkktoria/Flashcards.wkktoria/Program.cs
@@ -1,11 +1,11 @@
 ﻿using System.Configuration;
-using Flashcards.wkktoria;
 using Flashcards.wkktoria.Database;
 using Flashcards.wkktoria.Services;
+using Flashcards.wkktoria.UserInteractions;
 
 var databaseName = ConfigurationManager.AppSettings.Get("DatabaseName");
 var databasePassword = ConfigurationManager.AppSettings.Get("DatabasePassword");
-var connectionString = $@"Server=localhost,1433;User Id=sa;Password={databasePassword};TrustServerCertificate=true;";
+var connectionString = $"Server=localhost,1433;User Id=sa;Password={databasePassword};TrustServerCertificate=true;";
 
 var databaseManager = new DatabaseManager(connectionString, databaseName!);
 databaseManager.Initialize();
